Add ReceiveDamage script and calls for it to WeaponHitEmpower
Add ReceiveDamage script and calls for it in OnCollisionEnter method of
WeaponHitEmpower. ReceiveDamage has the health pool for limbs of the
character and the amount of overall health points which are reduced when
HitReceived method is called. Every limb part now has a tag to tell
where they are located within the body. When an enemy weapon hits any
limb part of the armature the victim's ReceiveDamage script is called
and health points of the right body parts get deducted. The velocity of
the weapon can determine how many health points are being reduced from
the victim's health pool but there aren't any numbers implemented yet.