Small physics changes
Small changes to physics, changed angle of the sword. Removed double
power mistake from sphere follower. Started trying to think how to make
a deadzone for SphereFollower so that if the crosshair is right behind
the character, the character will take another target from slightly to
side. The point is to prevent character from turning really slowly when
pointing directly behind